update: files (structure) / add: header component
