+++ .gitignore
@@ -69,3 +69,4 @@ yarn-error.log
 .yarn-integrity
 
 .vscode
+.editorconfig
