+++ src/components/layouts/layout.scss
@@ -1,0 +1,6 @@
+@import '../../scss/styles.scss';
+
+.header {
+  height: auto;
+  overflow: hidden;
+}
