+++ src/components/sections/dateheader/DateHeader.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Paragraph from '../../units/Paragraph';
+import './date.scss';
+
+const DateHeader = () => (
+  <div className="wrapper-animation">
+    <Paragraph className="wrapper-animation__item" inner="replace-image" />
+    <Paragraph className="wrapper-animation__item" inner="replace-image" />
+    <Paragraph className="wrapper-animation__item" inner="replace-image" />
+  </div>
+);
+
+export default DateHeader;
