+++ src/components/sections/dateheader/date.scss
@@ -1,0 +1,36 @@
+@import '../../../scss/styles.scss';
+
+.wrapper-animation {
+  display: flex;
+  width: 1500px;
+
+  @media (min-width: $breakpoint-md) {
+    width: 2500px; 
+  }
+
+  @media (min-width: $breakpoint-lg) {
+    width: 3000px; 
+  }
+}
+
+.wrapper-animation__item {
+animation: animate-image 10s linear infinite;
+background: url('../../../images/header.svg');
+background-repeat: no-repeat;
+background-size: 500px;
+height: 30px;
+margin: 10px 0 10px;
+text-indent: -9999px;
+transform: translate(-100%);
+width: 100%;
+
+@media (min-width: $breakpoint-md) {
+  background-size: 800px;
+  height: 40px;
+}
+
+@media (min-width: $breakpoint-lg) {
+  background-size: 1000px; 
+  height: 50px;
+}
+}
