+++ src/components/sections/herobanner/herobanner.scss
@@ -1,0 +1,3 @@
+.herobanner__video {
+  width: 100%;
+}
