+++ src/pages/proxy.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import Layout from '../components/layout';
 import SEO from '../components/seo';
 
 const Proxy = () => (
-  <Layout>
+  <>
     <SEO title="proxyOrg" />
     <h1>Proxy Organizacion 2019</h1>
-  </Layout>
+  </>
 );
 
 export default Proxy;
