+++ src/pages/sinestesia.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import Layout from '../components/layout';
 import SEO from '../components/seo';
 
 const Sinestesia = () => (
-  <Layout>
+  <>
     <SEO title="sinestesia" />
     <h1>Sinestesia</h1>
-  </Layout>
+  </>
 );
 
 export default Sinestesia;
